chore: move more files to schema_registry
- we move the schema reader and registry

- we rename the apis to controller and updates to the DI wiring
