+++ mypy.ini
@@ -15,7 +15,7 @@ warn_no_return = True
 warn_unreachable = True
 strict_equality = True
 
-[mypy-schema_registry.schema_registry_apis]
+[mypy-schema_registry.controller]
 ignore_errors = True
 
 [mypy-karapace.compatibility.jsonschema.checks]
