+++ tests/e2e/schema_registry/test_jsonschema.py
@@ -5,8 +5,8 @@ See LICENSE for details
 from jsonschema import Draft7Validator
 from karapace.client import Client
 from karapace.compatibility import CompatibilityModes
-from karapace.schema_reader import SchemaType
 from karapace.typing import SchemaMetadata, SchemaRuleSet
+from schema_registry.reader import SchemaType
 from tests.schemas.json_schemas import (
     A_DINT_B_DINT_OBJECT_SCHEMA,
     A_DINT_B_INT_OBJECT_SCHEMA,
