+++ tests/unit/schema_registry/__init__.py
@@ -1,0 +1,4 @@
+"""
+Copyright (c) 2024 Aiven Ltd
+See LICENSE for details
+"""
